wheEEEeEe
ok no entities atm. removed that script and scene, adjusted some things to help. pretty much everything can be a projectile or a weird projectile.

knockback works properly. It has its own gravity component for the curve instead of the weird thing i was doing beforehand. Now it keeps the curve, and the knockback velocity dissipates correctly.

knockback damping is no longer different for each character. Because that doesnt really make sense.

Added sub-air and sub-ground control accelerations.

if the horizontal speed is greater than the threshold value (for air or ground) then the acceleration can change to a different value. The larger value is also used when accelerating in the opposite direction of movement.

Discovered and fixed a bug where you could bypass the max ground speed by landing on the ground with a speed that is already greater.

Knockback Strength now means something; the distance that the hit character would move if they were at 0% and the force was applied to the x axis only.

hitboxes no longer check for specific types in their parents, and now just check for functions and parameters.

Fixed a bug in stages that could allow terrain rects to exist outside of the sprite rect.

Calculates camera_rect! fun times :)